Take event body and validate. Create utils file
