+++ src/lambda_src/lambda_handler.py
@@ -1,32 +1,37 @@
-import logging
 import json
-import os
-import requests
-
-from typing import Dict, List
+import logging
+from src.lambda_src.utils import get_openai_model_response, validate_request_body
 
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
 
-OPEN_AI_TOKEN = os.getenv("OPEN_AI_TOKEN")
-OPEN_AI_MODEL = os.getenv("OPEN_AI_MODEL")
-
 
 def lambda_handler(event, context):
-    # get header method and check if POST
-    # If not, return error
     http_method = event.get("http_method", "POST")
     if http_method.upper() != "POST":
+        logger.error(f"Invalid HTTP method used ({http_method})")
         return {
             "statusCode": 400,
             "headers": {
                 "Content-Type": "application/json"
             },
-            "body": json.dumps({"message":"Invalid HTTP method used"})
+            "body": "Invalid HTTP method used"
+        }
+
+    request_body = json.loads(event["body"])
+    is_valid_request_body = validate_request_body(request_body)
+    if not is_valid_request_body:
+        logger.error(f"Invalid request body ({request_body})")
+        return {
+            "statusCode": 400,
+            "headers": {
+                "Content-Type": "application/json"
+            },
+            "body": "Invalid request body. Must be in form [{image_path:'', image_captions:''}, {...}, ...]"
         }
 
     # Call function get response from GPT-4
-    initial_stories = get_openai_model_response()
+    initial_stories = get_openai_model_response(image_captions=request_body)
 
     return {
         "statusCode": 200,
@@ -35,51 +40,3 @@ def lambda_handler(event, context):
         },
         "body": initial_stories
     }
-
-
-def get_openai_model_response(image_captions: List[Dict[str, str]]):
-
-    url = "https://api.openai.com/v1/chat/completions"
-
-    system_content = "Given a list of dictionaries of photo captions from a vlog in the format " \
-                 " of this example: [{'image_path': 'birthday/BpsSOqpog98/" \
-                 "BpsSOqpog98-0190.jpg', 'image_caption': 'a child blowing out candles" \
-                 " at a birthday party'}, {...}, ...]" \
-                 ", create a vivid story " \
-                 "that incorporates the key elements from each " \
-                 "photo. Turn the photo descriptions into a fun and " \
-                 "engaging story. Your response should be a dictionary following the " \
-                 "format of this example: [{'image_path': 'birthday/BpsSOqpog98/" \
-                 "BpsSOqpog98-0190.jpg', 'image_caption': 'a child blowing out candles" \
-                 " at a birthday party', 'story': 'the story you generated'}, {...}, ...]"
-
-    user_content = json.dumps(image_captions)
-
-    headers = {
-        "Content-Type": "application/json",
-        "Authorization": f"Bearer {OPEN_AI_TOKEN}"
-    }
-
-    body = {
-        "model": OPEN_AI_MODEL,
-        "messages": [
-            {
-                "role": "system",
-                "content": system_content
-            },
-            {
-                "role": "user",
-                "content": user_content
-            }
-        ]
-    }
-
-    open_ai_response = requests.post(
-        url=url,
-        json=body,
-        headers=headers
-    ).json()
-
-    model_content = open_ai_response["choices"][0]["message"]["content"]
-
-    return json.loads(model_content)
